refactor(tui): split Model.Update by message ownership
Closes #732.

Update is now a short pipeline: clock tick, lifecycle finish messages,
global keys, overlay stack, then dispatchAppMsg. Confirm and choice
results live in app_update_dialog.go. Window size and contrast live in
app_update_layout.go. The remaining typed switch is dispatchAppMsg in
app_update_dispatch.go, grouped by the handler that owns each message.
